feat: 1:n (article, user)

+++ django/django_form/articles/views.py
@@ -40,24 +40,29 @@ def detail(request, pk):
     return render(request, 'articles/detail.html', context) 
 
 @require_POST
+@login_required
 def delete(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
+@login_required
 def update(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    if request.method == "POST":
-        form = ArticleForm(request.POST, instance=article)
+    if request.user == article.user:
+        if request.method == "POST":
+            form = ArticleForm(request.POST, instance=article)
 
-        if form.is_valid():
-            article = form.save(commit=False)
-            article.user = request.user
-            article.save()
-            return redirect('articles:detail', article.pk)
-    else:
-        form = ArticleForm(instance=article)
-    context = {
-        'form': form,
-    }
-    return render(request, 'articles/form.html', context)
+            if form.is_valid():
+                article = form.save(commit=False)
+                article.user = request.user
+                article.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
+        context = {
+            'form': form,
+        }
+        return render(request, 'articles/form.html', context)
+    return redirect('articles:index')
